Indicate a pre-render progress in the seekbar

+++ src/PlaybackController/PreRender.h
@@ -1,6 +1,6 @@
 /*
  * This file is part of sidplaywx, a GUI player for Commodore 64 SID music files.
- * Copyright (C) 2023 Jasmin Rutic (bytespiller@gmail.com)
+ * Copyright (C) 2023-2024 Jasmin Rutic (bytespiller@gmail.com)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -37,7 +37,7 @@ public:
 
 public:
 	int GetCurrentSongTimeMs() const;
-	int GetPreRenderProgressFactor() const;
+	double GetPreRenderProgressFactor() const;
 
 	void Stop();
 	void SeekTo(int timeMs, const SeekStatusCallback& callback);
